LUS-7476 kfilnd: Refactor kfilnd
Kfilnd was refactored to have an object oriented approach for the
three main objects: device, endpoint, and transaction. During this
refactoring, issues were discovered in the transaction state machine.
They revolved around buffer use-after-free issues. This lead to some
simplification of the transaction state machine.

In addition, kfilnd was based upon o2iblnd. Because of this, a lot of
unused/unimplemented code existed in kfilnd. Since this code served no
purpose, it was removed. This resulted in the startup of a kfilnd LNet
NI to be simplified.

Signed-off-by: Ian Ziemba <iziemba@cray.com>
Change-Id: If70309cff567581211003751f232629d84190078
Reviewed-on: https://es-gerrit.dev.cray.com/155300
Reviewed-by: Frank Zago <fzago@cray.com>