LUS-7470 kfilnd: Fix build and test issue
KFI LND was unable to build due to not setting the RMA RX context. In
addition, KFI LND was unable to run since the preferred RX context
was invalid. The CPT to NID LNet function may return a CPT that is not
assigned to the LNet NI. Using this CPT as the preferred RX context
may cause send operations to target a non-allocated RX context.

Signed-off-by: Ian Ziemba <iziemba@cray.com>
Change-Id: Ic2438bdb3efd23aed94fb7dc6705e531e10c8e43
Reviewed-on: https://es-gerrit.dev.cray.com/155291
Reviewed-by: Frank Zago <fzago@cray.com>