LUS-7484 kfilnd: Rework transaction remote MR keys
Remote memory region keys should be associated with a KFI LND domain
instead of being a global variable. This allows for MR keys to be
assigned differently for each KFI LND domain.

In addition, use a lock and unsigned int type for allocating keys.
This will avoid the undefined behavior of INT_MAX + 1 which may (or
may not) occur with atomic_t type.

Signed-off-by: Ian Ziemba <iziemba@cray.com>
Change-Id: I2aaee2636e6c289c0b241aa0147cb87fc10a82fa
Reviewed-on: https://es-gerrit.dev.cray.com/155319
Reviewed-by: Frank Zago <fzago@cray.com>