LUS-7489 kfilnd: Cleanup transaction code
When a transaction is complete, kfilnd_tn_finalize() is now called.
This function will cleanup transaction resources and call any
required LNet functions to complete the LNet operation.

In addition, unique fields are used to differentiate between a
transaction send and receive message. This makes the condition used to
free the send message buffer clear and not dependent on another
condition.

Signed-off-by: Ian Ziemba <iziemba@cray.com>
Change-Id: Ibbce79d1a218b85efc7144046d0c2db589fc642a
Reviewed-on: https://es-gerrit.dev.cray.com/155326
Reviewed-by: Frank Zago <fzago@cray.com>