LUS-7491 kfilnd: Fix setting of TN status
During transaction events, the event was not being progressed if an
invalid event occurs (which should not happen). In addition, when KFI
LND endpoint functions are called, the return code is automatically
assigned to the transaction status. This broke the CFS error injection
built into the LND.

To resolve these issues, during transaction event processing, return
codes which could impact the result of the transaction are only
applied when the transaction is being finalized. In addition, the
invalid events now apply an invalid return code to the transaction.

Signed-off-by: Ian Ziemba <iziemba@cray.com>
Change-Id: Ibc80d616ac324909f4afda498c2eb385741577fd
Reviewed-on: https://es-gerrit.dev.cray.com/155333
Reviewed-by: Frank Zago <fzago@cray.com>