LUS-7477 kfilnd: Implement a bulk response message
When a peer initiates a bulk put/get request, once the target has
completed the resulting RMA operation, the target sends a bulk
response with the result of the RMA operation.

When the peer sends a bulk request, it provides three fields: MR key,
cookie, and response RX context. The MR key is required for the target
to perform the RMA operation to move the data. The cookie is echoed
back to the initiator by the target in the bulk response. The cookie
is the pointer to the initiator's transaction structure. The response
RX context is required since the initiator locks the transaction to a
specific RX context.

In addition, cleanup of the transaction fields and redefining the
message operations a enums was done.

KFI LND no longer requires remote RMA events.

Signed-off-by: Ian Ziemba <iziemba@cray.com>
Change-Id: Icbfd807f24610c2ab8d152df5f2c8d534267a297
Reviewed-on: https://es-gerrit.dev.cray.com/155365
Reviewed-by: Frank Zago <fzago@cray.com>