LUS-7435 llite: release active extent on sync write commit
Processes can wait forever in osc_extent_wait() for the extent state
to change because the extent write is not started before the wait
begins. A kernel change to generic_write_sync() modified it to check
IOCB_DSYNC instead of O_SYNC. Thus an active extent is not released
(written) in osc_io_commit_async() in the synchronous case.

This mod is a version of the 2.12 commit
39c9408e4d1a63086fb84cb79efe54359af8e225

2.11 contains a patch for LU-8964, which moves some of the structs
needed by the current mod. The LU-8964 changes were subsequently
reverted (LU-11825) and thus the original patch from 2.7 applies
cleanly to 2.12.

Signed-off-by: Ann Koehler <amk@cray.com>
Change-Id: Iaf1784cd6b27d2211822f23a3db9e6dc3c2f584d
Reviewed-on: https://es-gerrit.dev.cray.com/155422
Reviewed-by: Andriy Skulysh <c17819@cray.com>
Tested-by: Elena Gryaznova <c17455@cray.com>
Reviewed-by: Chris Horn <hornc@cray.com>