LU-12639 tests: initialize variable sanity 317
sanity test 317 checks the file system type of $facet,
but facet is not initiaized in the test.

Replace the check of the facet file system type with the
variable ost1_FSTYPE.  The call to return after skip is
removed.

This is a backport to Lustre b2_12 of
Lustre-change: https://review.whamcloud.com/35716
Lustre-commit: c00be06a1f8f27eb5bd8bb47086d0f1e5b5f5f50

Fixes: 6115eb7fd55a ("LU-10370 ofd: truncate does not update blocks count on client")
Test-Parameters: trivial
Test-Parameters: fstype=zfs envdefinitions=ONLY=317 testlist=sanity

Signed-off-by: James Nunez <jnunez@whamcloud.com>
Change-Id: If67c4d786e4d23752effd1aaebc82bb1be8aceb5
Reviewed-by: Andreas Dilger <adilger@whamcloud.com>
Reviewed-by: Arshad Hussain <arshad.super@gmail.com>
Reviewed-on: https://review.whamcloud.com/36366
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>