LUS-7860 lnet: Add peer level aliveness information
Keep track of the aliveness of a peer so that we can optimize for
situations where an LNet router hasn't responded to a ping. In
this situation we consider all routes down, and we needn't spend time
inspecting each route, or inspecting all of the router's local and
remote interfaces in order to determine the router's aliveness.

Signed-off-by: Chris Horn <hornc@cray.com>
Change-Id: Ifc14c4f202a7d3d011245bafaaa51e5d9bcec4a9
Reviewed-on: https://es-gerrit.dev.cray.com/155863
Reviewed-by: Alexander Boyko <c17825@cray.com>
Reviewed-by: Alexey Lyashkov <c17817@cray.com>
Tested-by: Alexander Lezhoev <c17454@cray.com>