LU-11743 utils: allow lctl pool_list on separate MGS
Change lctl pool_list command to parse the configuration log directly
when run on a standalone MGS node.  This also allows the pool commands
to be run when only the MGS is started.

Also, those test scripts from the patch of LU-9899 to mount a client
on the standalone MGS to allow OST pools to work properly are cleared.

Signed-off-by: Emoly Liu <emoly@whamcloud.com>
Change-Id: I541765cb840b62edc4b99913aca68e50e972ee67