LUS-7979 mdt: add SEL to experimental features
Do not let SEL files to be created, SEL components to be added,
a template SEL layout to be created, a SEL layout to be inherited
if enable_experimental_features procfs flag is set.

Also, do not let a DOM layout to be inherited in this case as well
Also, as lod_verify_layout() has @is_from_disk flag, check for it
in both cases, as it is already created.

Signed-off-by: Vitaly Fertman <c17818@cray.com>
Change-Id: I64e48a4856635d55fe53a231c2858a7b3f6726b2
Reviewed-by: Alexey Lyashkov <c17817@cray.com>
Reviewed-by: Alexander Zarochentsev <c17826@cray.com>
Tested-by: Elena Gryaznova <c17455@cray.com>
Reviewed-on: https://es-gerrit.dev.cray.com/156063