LUS-8073 lnet: MR peers are not detected
The function lnet_is_peer_discovery_disabled() will return true if the
lnet_peer object passed as an argument does not have the MR feature
flag set in its lp_state field. But this function is used to determine
whether we should set this flag in the first place. Thus, MR peers are
not correctly identified.

Signed-off-by: Chris Horn <hornc@cray.com>
Change-Id: Ic24b017279f12db3c8b98628134afb1eebe4cb66
Reviewed-on: https://es-gerrit.dev.cray.com/156100
Reviewed-by: Alexander Boyko <c17825@cray.com>
Reviewed-by: Alexey Lyashkov <c17817@cray.com>
Reviewed-by: Andrew Perepechko <c17827@cray.com>
Tested-by: Elena Gryaznova <c17455@cray.com>