LUS-8141 kfilnd: Don't suppress send events
Suppressing send events complicates the tagged send and receive
implementation. Instead, do not suppress the send event of bulk
request messages.

Note that this does create an issue where the transaction asynchronous
event count does not get properly decremented if the bulk request
message fails. This is fixed with a tagged send and receive
implementation.

Signed-off-by: Ian Ziemba <iziemba@cray.com>
Change-Id: I27ea623b5f5697c8242fa982e833cccdd31ac0d9
Reviewed-on: https://es-gerrit.dev.cray.com/156194
Reviewed-by: Frank Zago <fzago@cray.com>