LUS-8142 kfilnd: Fix kfilnd locking
Kfilnd was using a spinlock to prevent concurrent access to the
transaction. This can create an issue where the thread goes to sleep
during memory region allocation with the spinlock held, and another
thread tries to process the transaction. To prevent this, use mutexes
to lock the transaction. Note that this should not be an issue since a
single core should be processing the transaction unless an error
occurs.

Signed-off-by: Ian Ziemba <iziemba@cray.com>
Change-Id: I6c7ead61e81c4aea940090ca0a95115a33015668
Reviewed-on: https://es-gerrit.dev.cray.com/156195
Reviewed-by: Frank Zago <fzago@cray.com>