LUS-8107 kfilnd: Run with RM disabled
Resource management mode relies on the flow control capabilities of
the provider. Instead, rely on the LNet credits to control kfilnd
flows.

In addition, when running with resource mode disabled, on transmit or
RMA error, the provider may internally retry. This behavior is not
desirable for kfilnd since a transmit or RMA error is a sign of a
transaction error.

With resource management mode disabled, queues need to be sized
appropriately. New module parameters are provided to configure the
size of the queues.

Signed-off-by: Ian Ziemba <iziemba@cray.com>
Change-Id: I2b67e9da925d218bc282f542c59b37d739d37219
Reviewed-on: https://es-gerrit.dev.cray.com/156268
Reviewed-by: Frank Zago <fzago@cray.com>