LUS-8108 kfilnd: Rework peer cache
Since the previous peer cache implementation did not take a reference
against the peer's KFI address when performing a transaction, it is
possible that the KFI address could have been removed from the KFI
address vector, but a transaction may still be using it.

To prevent a use-after-free issue, each transaction takes a reference
against the peer structure. Once the transaction is complete, a
reference is returned.

If a transaction completes with a non-zero return code, the peer
structure is marked for removal. Once marked, the next thread to
return a reference to the peer structure will also cause the peer
structure to be removed from the peer cache. Once all references are
zero, the peer structure will be freed. Future transaction to this
peer will require the allocation of a new peer structure.

In addition, the peer cache was changed to use the Linux provided hash
table.

Signed-off-by: Ian Ziemba <iziemba@cray.com>
Change-Id: I070ba4a9145570bd895f781ce368c3c9a924d1c5
Reviewed-on: https://es-gerrit.dev.cray.com/156269
Reviewed-by: Frank Zago <fzago@cray.com>