LUS-8109 kfilnd: Update kfilnd state machine
The kfilnd state machine has been refined with additional states and
events defined. Now, each time a transaction event occurs, the
transaction state will change.

The following states are introduced.
Wait Tagged Completion: The initiator or target is waiting for a
tagged send or receive to complete. Any event in this state will cause
the transaction to be finalized.

Fail: The initiator was performing a bulk transfer operation and an
error occurred while in the Register Memory or Wait Completion state.
While in the Fail state, another asynchronous event is required before
the transaction can be finalized.

In addition, Bulk RMA state was changed to Wait RMA Completion.

Signed-off-by: Ian Ziemba <iziemba@cray.com>
Change-Id: Iefd48860089d254fed5e5cd9fa46632483218ea5
Reviewed-on: https://es-gerrit.dev.cray.com/156270
Reviewed-by: Frank Zago <fzago@cray.com>