LUS-8111 kfilnd: Initial support for timeouts
Since bulk operations are dependent on the target to progress the
transaction, the initiator needs to associate a timeout with the
transaction to provide a time window in which the target must respond
by. Once the initiator has successfully sent the bulk request message
to the target, the initiator will transition from the Wait Completion
state to the Wait Tag Completion state. During this time, the
initiator enables a timeout timer to establish the time window the
target must respond within.

Once in the Wait Tag Completion state, three different events can
occur: TAG_RX_OK, TAG_RX_FAIL, and TIMEOUT. TAG_RX_OK occurs when the
target has responded to the bulk operation request. TAG_RX_FAIL occurs
when hardware was unable to post the tagged receive buffer. When
either of these occur, timeout timer cancellation is attempted. If
successful, the transaction is finalized. If not successful, the
transaction is changed from Wait Tag Completion to Wait Timeout
Completion. In this state, a TIMEOUT event will cause the transaction
to be finalized. Note that this TIMEOUT event will not cause the
transaction return code status to be set to timeout.

When in the Wait Tag Completion state, a TIMEOUT event causes the
initiator to cancel the tagged receive buffer. In addition, the
transaction state is changed to Wait Timeout Completion. In this
state, three events can occur: TAG_RX_OK, TAG_RX_FAIL, TAG_RX_CANCEL.
All of these events will cause the transaction to be finalized. Note
that if the event is TAG_RX_CANCEL, this means that the tagged receive
buffer was successfully cancelled due to the timeout. The transaction
return code status is updated to timeout to reflect this.

Signed-off-by: Ian Ziemba <iziemba@cray.com>
Change-Id: I2b5b2b751ae080a36de4151afc5ab2aa9aacfe69
Reviewed-on: https://es-gerrit.dev.cray.com/156284
Reviewed-by: Frank Zago <fzago@cray.com>