LUS-8211 kfilnd: Don't use flags for TN events
The TX_OK event is intended to only be used for post send operations.
Instead of overloading this event and using flags to make a decision,
define two new events, INIT_IMMEDIATE and INIT_BULK, and remove the
flags.

In addition, remove the remaining flag and add a bool argument for
whether or not a bulk buffer is a sink or source buffer.

Signed-off-by: Ian Ziemba <iziemba@cray.com>
Change-Id: I5d8b6cf2a2515b7f381614519c83821bd087b774
Reviewed-on: https://es-gerrit.dev.cray.com/156309
Reviewed-by: Frank Zago <fzago@cray.com>