LUS-8245 lnet: Convert health check error to CDEBUG
This error message shows up frequently with gnilnd because gnilnd
was never modified to support LNet health. The error condition is
harmless, so convert the message to a CDEBUG.

Note, this message was converted to a CDEBUG in master commit
18c850cb91a64fcc38ae0801e6fab983607ded71. However, we don't need
or want all of the changes from that commit in CLE 7.0.UP01.

Signed-off-by: Chris Horn <hornc@cray.com>
Change-Id: If9112fe77e02aaded401f6f26e4c967408ee8e3a