LUS-7700 tests: skip archive id in sanity-hsm/10a for external CDT
The external coordinator does not currently save the archive id of an
archived file in the MDT inode. The archive id is optionally provided
by the user in the archive request which is sent to other system
components for handling (e.g. a tiering or policy engine) which may
also determine the archive id used. The archive id is not sent to
the MDS when the archive operation takes place, so the external
coordinator has no knowledge of it and uses 0 as the archive id for
all archived files; the verification of the archive id in sanity-hsm
test 10a can be skipped when using the external coordinator, as it
will always fail.

Change-Id: Icf1e8750b550c4c53a256f8bebef32950e4d90c7
Signed-off-by: Nikitas Angelinas <nangelinas@cray.com>
Test-Parameters: testlist="sanity-hsm sanity-hsm-coordinator-external"
Reviewed-on: https://es-gerrit.dev.cray.com/156722
Reviewed-by: Nathan Rutman <nrutman@cray.com>
Reviewed-by: Sergey Cheremencev <c17829@cray.com>
Tested-by: Alexey Lyashkov <c17817@cray.com>