LUS-7999 libcfs: Cleanup libcfs_debug_msg use of snprintf
scnprintf returns the number of bytes written to the buffer.
snprintf returns the size of the buffer needed to satisfy
the request.

Prefer scnprintf when result is being used as the number
of bytes in a buffer.

Use snprintf when the result is used for sizing or resizing
a buffer.

Whamcloud-bug-id: LU-12861
Signed-off-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Change-Id: Ie65ea0f369f647d8b616cce797ea94f16e59f1b8
Reviewed-on: https://es-gerrit.dev.cray.com/156940
Reviewed-by: Alexander Boyko <c17825@cray.com>
Reviewed-by: Petros Koutoupis <pkoutoupis@cray.com>
Tested-by: Alexander Lezhoev <c17454@cray.com>