LU-12312 lnet: handle no discovery flag
When a peer is being discovered, it could indicate that its
discovery is disabled. In this case we shouldn't flag the
peer for deletion. This operation is reserved for cases when
the peer initiates a push to tell us that it has changed
its status from discovery to no discovery.

Fixes: 4577410165 ("LU-13028 lnet: advertise discovery when toggled")
Signed-off-by: Amir Shehata <ashehata@whamcloud.com>
Change-Id: I1d357555e234d828ba19b72fab1d041f93af548e
Reviewed-on: https://review.whamcloud.com/38229
Tested-by: jenkins <devops@whamcloud.com>
Reviewed-by: Serguei Smirnov <ssmirnov@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>
Reviewed-by: Sebastien Buisson <sbuisson@ddn.com>
Reviewed-by: Oleg Drokin <green@whamcloud.com>