LUS-8668 hsm: set archive id when using external coordinator
Set the archive id in the HSM xattr for each file before sending it to
the external system, instead of using the archive id in progress
messages from copytools, when using the external coordinator.

Since this allows the external coordinator to support archive ids,
enable the previously-disabled archive id verification in
sanity-hsm/10a.

Change-Id: I977081f9e3cacf333c0c300a87ec40aa4618f73c
Singed-off-by: Nikitas Angelinas <nangelinas@cray.com>
Reviewed-on: https://es-gerrit.dev.cray.com/157074
Reviewed-by: Andriy Skulysh <c17819@cray.com>
Reviewed-by: Nathan Rutman <nrutman@cray.com>
Tested-by: Jenkins Build User <nssreleng@cray.com>