LUS-8457 lnet: Primary NID of lolnd NID is the lolnd NID
We want Lustre traffic that is intended for the local peer to be sent
and received over the lolnd. The function ptlrpc_uuid_to_peer() will
currently resolve a NID to the lolnd NID, but ptlrpc_connection_get()
will overwrite this selection with the result from LNetPrimaryNID().

Have LNetPrimaryNID return the lolnd NID when it is passed the lolnd
NID.

Signed-off-by: Chris Horn <hornc@cray.com>
Change-Id: I9a2cd5587e82c4a4c188c4d0a788e211720a1cf1
Reviewed-on: https://es-gerrit.dev.cray.com/157137
Reviewed-by: Alexander Boyko <c17825@cray.com>
Reviewed-by: Alexey Lyashkov <c17817@cray.com>
Reviewed-by: Andrew Perepechko <c17827@cray.com>
Tested-by: Alexander Lezhoev <c17454@cray.com>