LUS-2616 hsm: save restore dfid in copytool instance of hai
When creating a volatile file for HSM restore operations, copy the fid of the
file created (dfid) to the copytool instance of the hai. When a restore
request is canceled the ioctl(LL_IOC_HSM_COPY_START) call in
llapi_hsm_action_begin() fails, which causes hcp to be freed and the dfid in its
copy of the hai lost. The final progress message sent to the CDT needs to
contain the correct dfid to pass checks in req_is_valid(); it uses the dfid from
the copytool's instance of the hai in the cancellation case, so this needs to
contain the correct value.

Change-Id: If35698315616b1da44491897fae032db0a6acf68
Signed-off-by: Nikitas Angelinas <nangelinas@cray.com>
Test-Parameters: testlist="sanity-hsm sanity-hsm-coordinator-external"
Reviewed-on: https://es-gerrit.dev.cray.com/157198
Reviewed-by: Nathan Rutman <nrutman@cray.com>
Tested-by: Alexander Lezhoev <c17454@cray.com>
Reviewed-by: Sergey Cheremencev <c17829@cray.com>