LUS-8294 osd-zfs: Limit failures in lsnapshot test_4
The current lsnapshot test bounces the mgs and waits
for the initial number of connections to match.

This test fails due to connections not being established or
additional connections adding after the mgs is bounced.

The connection count mismatches are not fatal so don't
fail when the occur.

In addition an lsnapshot may occur before the lustre
file system is fully restored. The snapshot will normally succeed
with an additional small delay, so retry the snapshot in such cases.

Test-Parameters: fstype=zfs
Signed-off-by: Shaun Tancheff <shaun.tancheff@hpe.com>
Change-Id: I43b916201bb377c7d8967288bef11d2ba29d544e
Reviewed-on: https://es-gerrit.dev.cray.com/157227
Reviewed-by: Petros Koutoupis <pkoutoupis@cray.com>
Reviewed-by: Elena Gryaznova <elena.gryaznova@hpe.com>
Tested-by: Elena Gryaznova <elena.gryaznova@hpe.com>