LUS-8841 tests: data corruption due to RPC reordering
This commit adds a new test case. The client
can receive a reply from from the original RPC
or the resent one, clearing PG_Writeback and
making it possible to modify the same page on
the client. However, the resent RPC or the
original RPC can proceed after this modification.
Eventually, stale data gets to disk.

Change-Id: Ie84ee7504209c3df6acdf71fbe8c25b2034988c2
Signed-off-by: Andrew Perepechko <andrew.perepechko@hpe.com>
Test-Parameters: testlist=recovery-small/135b
Reviewed-on: https://es-gerrit.dev.cray.com/157259
Reviewed-by: Alexander Zarochentsev <c17826@cray.com>
Tested-by: Alexander Lezhoev <c17454@cray.com>
Reviewed-by: Vitaly Fertman <c17818@cray.com>