LUS-8635 ldlm: ldlm_flock_deadlock() ASSERTION(req != lock) failed
A client gets evicted and reconnects, so there can be a window
with 2 exports with flocks from the same client.
In this case during deadlock search from the new export we can go
to the old export and back to the new one.

Failed exports should be excluded from
deadlock search.

Change-Id: I4bd078b793c41d7b79d55396ecd6dc7f0bdc1623
Signed-off-by: Andriy Skulysh <c17819@cray.com>
Reviewed-on: https://es-gerrit.dev.cray.com/157284
Reviewed-by: Alexander Boyko <c17825@cray.com>
Reviewed-by: Vitaly Fertman <c17818@cray.com>
Tested-by: Alexander Lezhoev <c17454@cray.com>