LUS-8908 kfilnd: Fix setting of transaction buffer
The kfabric data transfer APIs do not allow for an offset to be
applied to buffer. Instead, the kfabric API expects the buffer to have
any user defined offset factored into it. Kfilnd was not taking this
into account. When LNet was passing kfilnd an IOV, this IOV was being
passed into the kfabric API without taking into consider the offset
and length of the IOV. This resulted in kfabric RMA operations with an
incorrect size.

To resolve this issue, kfilnd_tn_set_buf() is defined. This function
will take the LNet IOV, apply the LNet offset and length, and store
the resulting IOV within the transaction structure. This new IOV will
be used for kfabric memory registration and data transfers.

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>
Change-Id: Ied7929bf11255acaf72921a5a49e06ee76772c4f
Reviewed-on: https://es-gerrit.dev.cray.com/157404
Reviewed-by: Frank Zago <fzago@cray.com>
Tested-by: Alexander Lezhoev <c17454@cray.com>
Reviewed-by: Alexander Lezhoev <c17454@cray.com>