LUS-8520 ptlrpc: decrease time between reconnection
When a connection get a timeout or get an error reply from a sever,
the next attempt happens after PING_INTERVAL. It is equal to
obd_timeout/4. When a first reconnection fails, a second go to
failover pair. And a third connection go to a original server.
Only 3 reconnection before server evicts client base on blocking
ast timeout. Some times a first failed and the last is a bit late,
so client is evicted. It is better to try reconnect with a timeout
equal to a connection request deadline, it would inrceases a number
of attempts in 5 times for large obd_timeout.
For example,
obd_timeout=200
 - [ 1597902357, CONNECTING ]
 - [ 1597902357, FULL ]
 - [ 1597902422, DISCONN ]
 - [ 1597902422, CONNECTING ]
 - [ 1597902433, DISCONN ]
 - [ 1597902473, CONNECTING ]
 - [ 1597902473, DISCONN ] <- ENODEV from failover pair
 - [ 1597902523, CONNECTING ]
 - [ 1597902539, DISCONN ]

The patch adds a logic to wakeup pinger for failed connection request
with ETIMEDOUT or ENODEV. It adds imp_next_ping processing for
ptlrpc_pinger_main time_to_next_wake calculation, and fix setting
of imp_next_ping value.

Signed-off-by: Alexander Boyko <alexander.boyko@hpe.com>
Change-Id: I0bd519821de4d61f137e674afa5f71f605ed8632
Reviewed-on: https://es-gerrit.dev.cray.com/157718
Reviewed-by: Vitaly Fertman <c17818@cray.com>
Reviewed-by: Alexey Lyashkov <c17817@cray.com>
Tested-by: Elena Gryaznova <c17455@cray.com>