LUS-9109 lnet: Age peer NI out of recovery
No longer send recovery pings to a peer NI that has been in recovery
for the recovery time limit. A peer NI will become eligible for
recovery again once we receive a message from it.

The existing lpni_last_alive field is utilized for this new purpose.
Current uses are converted from ktime_get_seconds() to
ktime_get_real_seconds().

A check for NULL lpni is removed from
lnet_handle_remote_failure_locked() because all callers of that
function already ensure the lpni is non-NULL.

Amongst other changes, lnet_peer_ni_add_to_recoveryq_locked() is
modified to take a ref on the peer NI if it is added to the recovery
queue. Previously, it was the responsibility of callers to take this
ref.

Test-Parameters: trivial
Signed-off-by: Chris Horn <chris.horn@hpe.com>
Change-Id: Ib4676540ac4bb040690a4fb047236c54eea0e752
Reviewed-on: https://es-gerrit.dev.cray.com/157734
Reviewed-by: Alexander Boyko <c17825@cray.com>
Reviewed-by: Alexey Lyashkov <c17817@cray.com>
Tested-by: Vitaly Fertman <c17818@cray.com>