LUS-9343 kfilnd: Fix checking of peer field
The transaction debug and error print macros only check the
transaction peer field if it is NULL or not. There is a third state
the peer field can be which is error. Define the KFILND_TN_PEER_VALID
macro which can be used to verify the peer field is valid or not. It
accounts for all three states.

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>
Change-Id: Ibe2d51067db6dcf094eb3aae69a6a9aaef0f5d82
Reviewed-on: https://es-gerrit.dev.cray.com/157850
Reviewed-by: Frank Zago <fzago@cray.com>
Tested-by: Ian Ziemba <iziemba@cray.com>
Reviewed-by: Ian Ziemba <iziemba@cray.com>