LUS-9288 kfilnd: Fix kfilnd threading
The current thread work pool implementation was unable to scale when a
high rate of events occurred. The end result was all worker threads
were consumed and new events were dropped since they were unable to be
processed.

The new approach allocates a "pool" of work items for each CQ and EQ.
When an event occurs, the pool of work items are scheduled on a kfilnd
work queue. For the CQ, the pool work items are mapped to cores in the
corresponding CPT. For the CQ, the end result is multiple work queue
threads within the CPT processing the CQ until it is empty. For the
EQ, since no CPT is assigned to the EQ, the EQ thread is not assigned
to a specific core.

Signed-off-by: Ian Ziemba <ian.ziemba@hpe.com>
Change-Id: Ia8060a84dd5d1e0ebbe22127c6fd4ed0ea8ffb64
Reviewed-on: https://es-gerrit.dev.cray.com/157862
Reviewed-by: Frank Zago <fzago@cray.com>
Tested-by: Ian Ziemba <iziemba@cray.com>
Reviewed-by: Ian Ziemba <iziemba@cray.com>