LUS-9202 tests: race on lu_device_init
On concurrent lu_device_init()-s it is possible that a process which
skips lu device type initialization reaches keys_fill() before
LCT_QUIESCENT flag of a key is cleared. Necessary key values do not
get initialized in context which leads to crash.

Test to illustrate the case is added.

Signed-off-by: Vladimir Saveliev <c17830@cray.com>
Change-Id: I0983b5a5d7c1548cb4aa0787759c7f8ee6e375aa
Test-Parameters: testlist=insanity/15
Reviewed-on: https://es-gerrit.dev.cray.com/#/c/157846
Reviewed-by: Alexander Boyko <c17825@cray.com>
Reviewed-by: Elena Gryaznova <c17455@cray.com>
Tested-by: Elena Gryaznova <c17455@cray.com>